+++ apps/opentelemetry/test/opentelemetry_SUITE.erl
@@ -28,7 +28,7 @@ all() ->
 
 all_cases() ->
     [with_span, macros, child_spans,
-     update_span_data, tracer_instrumentation_library, tracer_previous_ctx, stop_temporary_app,
+     update_span_data, tracer_instrumentation_scope, tracer_previous_ctx, stop_temporary_app,
      reset_after, attach_ctx, default_sampler, non_recording_ets_table,
      root_span_sampling_always_on, root_span_sampling_always_off,
      record_but_not_sample, record_exception_works, record_exception_with_message_works,
@@ -103,7 +103,7 @@ init_per_testcase(too_many_attributes, Config) ->
     Tid = ets:new(exported_spans, [public, bag]),
     otel_batch_processor:set_exporter(otel_exporter_tab, Tid),
     [{tid, Tid} | Config];
-init_per_testcase(tracer_instrumentation_library, Config) ->
+init_per_testcase(tracer_instrumentation_scope, Config) ->
     application:set_env(opentelemetry, processors, [{otel_batch_processor, #{scheduled_delay_ms => 1}}]),
     {ok, _} = application:ensure_all_started(opentelemetry),
     %% adds an exporter for a new table
@@ -139,34 +139,34 @@ end_per_testcase(_, _Config) ->
     ok.
 
 disable_auto_creation(_Config) ->
-    {_, #tracer{instrumentation_library=Library}} = opentelemetry:get_tracer(
+    {_, #tracer{instrumentation_scope=Library}} = opentelemetry:get_tracer(
                                                       opentelemetry:get_application(kernel)),
     ?assertEqual(undefined, Library),
     ok.
 
 old_disable_auto_creation(_Config) ->
-    {_, #tracer{instrumentation_library=Library}} = opentelemetry:get_tracer(
+    {_, #tracer{instrumentation_scope=Library}} = opentelemetry:get_tracer(
                                                       opentelemetry:get_application(kernel)),
     ?assertEqual(undefined, Library),
     ok.
 
 application_tracers(_Config) ->
-    {_, #tracer{instrumentation_library=Library}} = opentelemetry:get_tracer(
+    {_, #tracer{instrumentation_scope=Library}} = opentelemetry:get_tracer(
                                                       opentelemetry:get_application(kernel)),
-    ?assertEqual(<<"kernel">>, Library#instrumentation_library.name),
+    ?assertEqual(<<"kernel">>, Library#instrumentation_scope.name),
 
     %% tracers are unique by name/version/schema_url
     NewKernelTracer = opentelemetry:get_tracer(kernel, <<"fake-version">>, undefined),
-    {_, #tracer{instrumentation_library=NewLibrary}} = NewKernelTracer,
-    ?assertEqual(<<"kernel">>, NewLibrary#instrumentation_library.name),
-    ?assertEqual(<<"fake-version">>, NewLibrary#instrumentation_library.version),
-    ?assertEqual(undefined, NewLibrary#instrumentation_library.schema_url),
+    {_, #tracer{instrumentation_scope=NewLibrary}} = NewKernelTracer,
+    ?assertEqual(<<"kernel">>, NewLibrary#instrumentation_scope.name),
+    ?assertEqual(<<"fake-version">>, NewLibrary#instrumentation_scope.version),
+    ?assertEqual(undefined, NewLibrary#instrumentation_scope.schema_url),
 
     Tracer = opentelemetry:get_tracer(kernel, <<"fake-version">>, <<"http://schema.org/myschema">>),
-    {_, #tracer{instrumentation_library=NewLibrary1}} = Tracer,
-    ?assertEqual(<<"kernel">>, NewLibrary1#instrumentation_library.name),
-    ?assertEqual(<<"fake-version">>, NewLibrary1#instrumentation_library.version),
-    ?assertEqual(<<"http://schema.org/myschema">>, NewLibrary1#instrumentation_library.schema_url),
+    {_, #tracer{instrumentation_scope=NewLibrary1}} = Tracer,
+    ?assertEqual(<<"kernel">>, NewLibrary1#instrumentation_scope.name),
+    ?assertEqual(<<"fake-version">>, NewLibrary1#instrumentation_scope.version),
+    ?assertEqual(<<"http://schema.org/myschema">>, NewLibrary1#instrumentation_scope.schema_url),
 
     ok.
 
@@ -434,15 +434,15 @@ update_span_data(Config) ->
     ok.
 
 
-tracer_instrumentation_library(Config) ->
+tracer_instrumentation_scope(Config) ->
     Tid = ?config(tid, Config),
 
     TracerName = tracer1,
     TracerVsn = <<"1.0.0">>,
-    Tracer = {_, #tracer{instrumentation_library=IL}} =
+    Tracer = {_, #tracer{instrumentation_scope=IL}} =
         opentelemetry:get_tracer(TracerName, TracerVsn, "http://schema.org/myschema"),
 
-    ?assertMatch({instrumentation_library,<<"tracer1">>,<<"1.0.0">>,<<"http://schema.org/myschema">>},
+    ?assertMatch({instrumentation_scope,<<"tracer1">>,<<"1.0.0">>,<<"http://schema.org/myschema">>},
                  IL),
 
     SpanCtx1 = otel_tracer:start_span(Tracer, <<"span-1">>, #{}),
@@ -451,8 +451,8 @@ tracer_instrumentation_library(Config) ->
 
     [Span1] = assert_exported(Tid, SpanCtx1),
 
-    ?assertMatch({instrumentation_library,<<"tracer1">>,<<"1.0.0">>,<<"http://schema.org/myschema">>},
-                 Span1#span.instrumentation_library).
+    ?assertMatch({instrumentation_scope,<<"tracer1">>,<<"1.0.0">>,<<"http://schema.org/myschema">>},
+                 Span1#span.instrumentation_scope).
 
 %% check that ending a span results in the tracer setting the previous tracer context
 %% as the current active and not use the parent span ctx of the span being ended --